Sample data: health data from the European Social Survey Wave 7

This dataset contains data from the ESS Wave 7 (2014) for Germany, the UK and France that describes respondents physical and mental health as well as specific health behaviours. It can be used to demonstrate most functions of this package.

Format

A dataframe with 7,226 rows and 23 columns

cntry

Country

gndr

Gender (1 = male, 2 = female)

agea

Age

eisced

Education

pweight

Population weight - adjust for different population sizes between countries

pspwght

Post-stratification weight - adjusts for bias in national samples due to sampling error and non-response (considering gender, age, education and region) AND for unequal inclusion probabilities due to survey design

health

Overall health, from 1 = very good to 5 = very bad

height

Height (cm)

weight

Weight (kg)

icbrnct

Respondent born in country, 1 = yes, 2 = no

etfruit

How often eat fruit, excluding drinking juice, from 1 = three times or more per day to 7 = never

eatveg

How often eat vegetables or salad, excluding potatoes, from 1 = three times or more per day to 7 = never

dosprt

Do sports or other physical activity, how many of last 7 days

cgtsmke

Cigarettes smoking behaviour, from 1 = I smoke daily to 5 = I have never smoked

alcfreq

How often drink alcohol, from 1 = every day to 5 = never

fltdpr

Felt depressed in past week, from 1 = none/almost none of the time to 4 = all/almost all of the time (same for next 7 variables)

flteeff

Felt everything did as effort

slprl

Sleep was restless

wrhpp

Were happy

fltlnl

Felt lonely

enjlf

Enjoyed life

fltsd

Felt sad

cldgng

Could not get going

References

https://www.europeansocialsurvey.org/data/download.html?r=7

Author

ESS-ERIC, selected by package author