Glance a multiple imputation mice pooled object

Note that the mice authors prefer to tidy mipo rather than mira objects and have now included tidy.mipo and glance.mipo into their package. The mira functions here are mostly retained for compatibility with my earlier code.

# S3 method for class 'mira'
glance(x, adj_R2 = TRUE, ...)

Arguments

x

An object with multiply-imputed models from mice (class: mira)

adj_R2

Should an adjusted R2 be returned as well? Only applicable if x contains lm-models

...

extra arguments (not used)

Value

a tibble with one row

Note

If x contains lm models, R2 is included in the output

Examples

if (FALSE) { # \dontrun{
library(mice)
data <- airquality
data[4:10, 3] <- rep(NA, 7)
data[1:5, 4] <- NA
tmp <- mice(data, m = 5, seed = 500, printFlag = FALSE)
mod <- with(tmp, lm(Ozone ~ Solar.R + Wind))
glance(mod)
} # }