Plot mediation model with one or more mediators

Returns graphViz code and graph for (multiple) mediation model. Coefficients and significance values for paths need to be provided, based on the format returned by `run_moderation()`

plot_mediation(
  IV,
  DV,
  Ms,
  data,
  digits = 2,
  coef_offset = length(Ms),
  filename = NULL,
  ind_p_values = FALSE
)

Arguments

IV

Character. Name of predictor

DV

Character. Name of dependent variable

Ms

Character vector. Names of mediator variables

data

Dataframe with coefficients and significance values. See details.

digits

Number of digits for rounding

coef_offset

Tibble with values to position mediators. If not provided, function will align mediators automatically, which is unlikely to provide a well-aligned path (except for cases when offset has been implemented for that number of mediators, currently 1 and 3). However, returned code can still be edited. See timesaveR:::.coef_offset_3 for an example of an offset tibble.

filename

If provided, graph will be saved as .svg file.

ind_p_values

Should significance stars be shown for indirect effects, based on pvalues passed in data? If FALSE, indirect effects with confidence intervals that do not include zero are bolded

Value

A list of a the graph and the associated code.

Examples

if (FALSE) {
# Estimate mediation model

res <- run_mediation(ess_health, fltdpr, health, dosprt, agea, bootstraps = 100)

# Run plot command
plot_mediation(
  IV = "Frequency of  <br /> feeling depressed",
  DV = "Self-reported <br /> poor health", Ms = "Frequency of <br /> physical activity", data = res
)
}