Create a summary table comparing standardized and non-standardized linear models

This function creates a summary table for lm models (including mice::mira objects containing lm-models) that shows a standardized and non-standardized version of the model side-by-side. Several pairs of such models can be compared side-by-side.

report_lm_with_std(
  mod,
  mod_std,
  conf_level = 0.95,
  coef_renames = NULL,
  fmt = "%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  R2_change = FALSE,
  notes = list(NULL),
  apa_style = TRUE,
  stars = std_stars_pad,
  ...
)

Arguments

mod

A lm-model/mira object of lm models, with variables not standardized (or a list of such models)

mod_std

A lm-model/mira object of lm models, with standardized variables. Can be created with lm_std (or a list of such models)

conf_level

Confidence level to use for confidence intervals, defaults to .95

coef_renames

A named character vector with new names for the coefficients or a tibble as provided by get_coef_rename_tribble for variables. If NULL, then the coefficients are not renamed.

fmt

determines how to format numeric values

integer: the number of digits to keep after the period format(round(x, fmt), nsmall=fmt)

character: passed to the sprintf function (e.g., '%.3f' keeps 3 digits with trailing zero). See ?sprintf

function: returns a formatted character string.

LaTeX output: To ensure proper typography, all numeric entries are enclosed in the \num{} command, which requires the siunitx package to be loaded in the LaTeX preamble. This behavior can be altered with global options. See the 'Details' section.

statistic_vertical

Should standard errors and CIs be shown below coefficients? Defaults to horizontal layout

filename

the file name to create a HTML file on disk.

model_names

If several pairs of models are to be plotted side by side, indicate the label for each *pair* here

show_nimp

Logical - DEFUNCT. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic

R2_change

Logical. Report R2 change and F-test to compare models.

notes

List of notes to append to bottom of table. An explanation of significance stars is automatically added. If the std models were run with a helper function in this package, a note regarding the standardization is also automatically added.

apa_style

Logical, should APA-style formatting be applied

stars

Named vector of significance stars and their thresholds, check `timesaveR:::std_stars_pad` for default.

...

Arguments passed on to modelsummary::modelsummary

coef_map

character vector. Subset, rename, and reorder coefficients. Coefficients omitted from this vector are omitted from the table. The order of the vector determines the order of the table. coef_map can be a named or an unnamed character vector (see the Examples section below). If coef_map is a named vector, its values define the labels that must appear in the table, and its names identify the original term names stored in the model object: c("hp:mpg"="HPxM/G").

coef_omit

string regular expression. Omits all matching coefficients from the table using grepl(perl=TRUE). This argument uses perl-compatible regular expressions, which allows expressions such as "Int|ABC" which omits coefficients matching either "Int" or "ABC", and "^(?!.*Intercept)"` which omits every term except the intercept.

coef_rename

named character vector or function which returns a named vector. Values of the vector refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term. If you provide a function to coef_rename, modelsummary will create a named vector for you by deriving the new variable names from the vector of original term names with your function.

Value

A list with `gt_tab` (the gt-table object including the parts of the table that can be created with gt. This can be post-processed and formatted with functions in the gt-package, but does not include the lower part with model statistics, e.g., R^2.) and `html_code` (the code that creates the full table, and is used to render it in the Viewer).

Examples

# Standard lm model
mod1 <- lm(mpg ~ hp + wt, mtcars)

# Model with standardized coefficients

mod2 <- lm_std(mpg ~ hp + wt, mtcars)

report_lm_with_std(mod1, mod2)