Create a summary table comparing standardized and non-standardized proportional odd logistic regression models

This function creates a summary table for polr models (including mice::mira objects containing polr-models) that shows a standardized and non-standardized version of the model side-by-side. Several pairs of such models can be compared side-by-side.

report_polr_with_std(
  mod,
  mod_std,
  OR = TRUE,
  conf_level = 0.95,
  fmt = "%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  notes = list(),
  apa_style = TRUE,
  stars = std_stars_pad,
  ...
)

Arguments

mod

A polr-model/mira object of polr models, with variables not standardized (or a list of such models)

mod_std

A polr-model/mira object of polr models, with standardized predictor variables (or a list of such models)

OR

Logical. Should odds ratios be shown instead of typical coefficients. If TRUE, estimates are exponentiated

conf_level

Confidence level to use for confidence intervals, defaults to .95

fmt

determines how to format numeric values

integer: the number of digits to keep after the period format(round(x, fmt), nsmall=fmt)

character: passed to the sprintf function (e.g., '%.3f' keeps 3 digits with trailing zero). See ?sprintf

function: returns a formatted character string.

LaTeX output: To ensure proper typography, all numeric entries are enclosed in the \num{} command, which requires the siunitx package to be loaded in the LaTeX preamble. This behavior can be altered with global options. See the 'Details' section.

statistic_vertical

Should standard errors and CIs be shown below coefficients? Defaults to horizontal layout

filename

the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)

model_names

If several pairs of models are to be plotted side by side, indicate the label for each *pair* here

show_nimp

Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic

notes

List of notes to append to bottom of table. An explanation of significance stars is automatically added. A note is also added stating that dummy variables were not scaled in standardization. If you approached standardisation differently, that should be removed.

apa_style

Logical, should APA-style formatting be applied

stars

Named vector of significance stars and their thresholds, check `timesaveR:::std_stars_pad` for default.

...

Arguments passed on to modelsummary::modelsummary

output

filename or object type (character string)

Supported filename extensions: .docx, .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "gt", "kableExtra", "huxtable", "flextable", "jupyter". The "modelsummary_list" value produces a lightweight object which can be saved and fed back to the modelsummary function.

Warning: Users should not supply a file name to the output argument if they intend to customize the table with external packages. See the 'Details' section.

LaTeX compilation requires the booktabs and siunitx packages, but siunitx can be disabled or replaced with global options. See the 'Details' section.

The default output formats and table-making packages can be modified with global options. See the 'Details' section.

estimate

string or glue string of the estimate to display (or a vector with one string per model). Valid entries include any column name of the data.frame produced by get_estimates(model). Examples:

"estimate"

"{estimate} ({std.error}){stars}"

"{estimate} [{conf.low}, {conf.high}]"

vcov

robust standard errors and other manual statistics. The vcov argument accepts six types of input (see the 'Details' and 'Examples' sections below):

NULL returns the default uncertainty estimates of the model object

string, vector, or (named) list of strings. "iid", "classical", and "constant" are aliases for NULL, which returns the model's default uncertainty estimates. The strings "HC", "HC0", "HC1" (alias: "stata"), "HC2", "HC3" (alias: "robust"), "HC4", "HC4m", "HC5", "HAC", "NeweyWest", "Andrews", "panel-corrected", "outer-product", and "weave" use variance-covariance matrices computed using functions from the sandwich package, or equivalent method. The behavior of those functions can (and sometimes must) be altered by passing arguments to sandwich directly from modelsummary through the ellipsis (...), but it is safer to define your own custom functions as described in the next bullet.

function or (named) list of functions which return variance-covariance matrices with row and column names equal to the names of your coefficient estimates (e.g., stats::vcov, sandwich::vcovHC, function(x) vcovPC(x, cluster="country")).

formula or (named) list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).

named list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a named list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below. Warning: since this list of vectors can include arbitrary strings or numbers, modelsummary cannot automatically calculate p values. The stars argument may thus use incorrect significance thresholds when vcov is a list of vectors.

coef_map

character vector. Subset, rename, and reorder coefficients. Coefficients omitted from this vector are omitted from the table. The order of the vector determines the order of the table. coef_map can be a named or an unnamed character vector (see the Examples section below). If coef_map is a named vector, its values define the labels that must appear in the table, and its names identify the original term names stored in the model object: c("hp:mpg"="HPxM/G").

coef_omit

string regular expression. Omits all matching coefficients from the table using grepl(perl=TRUE). This argument uses perl-compatible regular expressions, which allows expressions such as "Int|ABC" which omits coefficients matching either "Int" or "ABC", and "^(?!.*Intercept)"` which omits every term except the intercept.

coef_rename

named character vector or function which returns a named vector. Values of the vector refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term. If you provide a function to coef_rename, modelsummary will create a named vector for you by deriving the new variable names from the vector of original term names with your function.

gof_map

rename, reorder, and omit goodness-of-fit statistics and other model information. This argument accepts 3 types of values:

NULL (default): the modelsummary::gof_map dictionary is used for formatting, and all unknown statistic are included.

data.frame with 3 columns named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples' section below.

list of lists, each of which includes 3 elements named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples section below'.

gof_omit

string regular expression. Omits all matching gof statistics from the table. This argument uses perl-compatible regular expressions (grepl(perl=TRUE)), which allows expressions such as ".*" which omits everything, and "^(?!R2|Num)" which omits every term except those that start with "R2" or "Num".

group

transpose or group models and estimates (formula). The left side of the formula represents rows and the right side columns.

Formula with two components called term and model.

term ~ model

model ~ term

Formula with three components, including one of the column names in the data frame produced by get_estimates(model). Examples:

response + term ~ model

term ~ model + y.level

group_map

named or unnamed character vector. Subset, rename, and reorder coefficient groups specified in the group argument. See coef_map.

add_rows

a data.frame (or tibble) with the same number of columns as your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the row positions. See Examples section below.

align

A string with a number of characters equal to the number of columns in the table (e.g., align = "lcc"). Valid characters: l, c, r, S.

"l": left-aligned column

"c": centered column

"r": right-aligned column

"d": dot-aligned column. Only supported for LaTeX/PDF tables produced by kableExtra. These commands must appear in the LaTeX preamble (they are added automatically when compiling Rmarkdown documents to PDF):

\usepackage{booktabs}

\usepackage{siunitx}

\newcolumntype{d}{S[input-symbols = ()]}

title

string

escape

boolean TRUE escapes or substitutes LaTeX/HTML characters which could prevent the file from compiling/displaying. This setting does not affect captions or notes.

Details

The R2 shown is the maximum likelihood pseudo R2 returned by pR2.

Examples

library(MASS)
#> 
#> Attaching package: ‘MASS’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
pov_att <- polr(poverty ~ religion + age + gender, data = WVS)
pov_att_std <- polr_std(poverty ~ religion + age + gender, data = WVS)
report_polr_with_std(pov_att, pov_att_std, coef_omit = "\\|")
#> 
#> Re-fitting to get Hessian
#> fitting null model for pseudo-r2
#> $gt_tab
#> <div id="gynoykymid" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: times, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #gynoykymid .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: normal;
#>   font-style: none;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_title {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #gynoykymid .gt_subtitle {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #gynoykymid .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #gynoykymid .gt_col_headings {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #gynoykymid .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #gynoykymid .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #gynoykymid .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #gynoykymid .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #gynoykymid .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #gynoykymid .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #gynoykymid .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #gynoykymid .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #gynoykymid .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: none;
#>   border-top-width: 1px;
#>   border-top-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #gynoykymid .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #gynoykymid .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #gynoykymid .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #gynoykymid .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #gynoykymid .gt_first_summary_row {
#>   border-top-style: none;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #gynoykymid .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #gynoykymid .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: none;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #gynoykymid .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #gynoykymid .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #gynoykymid .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #gynoykymid .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #gynoykymid .gt_left {
#>   text-align: left;
#> }
#> 
#> #gynoykymid .gt_center {
#>   text-align: center;
#> }
#> 
#> #gynoykymid .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #gynoykymid .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #gynoykymid .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #gynoykymid .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #gynoykymid .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #gynoykymid .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #gynoykymid .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #gynoykymid .gt_slash_mark {
#>   font-size: 0.7em;
#>   line-height: 0.7em;
#>   vertical-align: 0.15em;
#> }
#> 
#> #gynoykymid .gt_fraction_numerator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: 0.45em;
#> }
#> 
#> #gynoykymid .gt_fraction_denominator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: -0.05em;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1"><em><center>OR [95% CI]</center></em></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1"><em><center>Stand. OR [95% CI]</center></em></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_left"><div class='gt_from_md'><p>religionyes</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>0.95     [0.82, 1.09]</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>0.95     [0.82, 1.09]</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_left"><div class='gt_from_md'><p>age</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.01 *** [1.01, 1.02]</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.25 *** [1.19, 1.32]</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_left"><div class='gt_from_md'><p>gendermale</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.16 **  [1.05, 1.29]</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.16 **  [1.05, 1.29]</p>
#> </div></td></tr>
#>   </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3">Given that dummy variables lose their interpretability when standardized, only continuous predictors are standardized.</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3">† <em>p</em> &lt; .1, * <em>p</em> &lt; .05, ** <em>p</em> &lt; .01, *** <em>p</em> &lt; .001</td>
#>     </tr>
#>   </tfoot>
#>   
#> </table>
#> </div>
#> 
#> $html_code
#> [1] "<!DOCTYPE html>\n<html lang=\"en\">\n<head>\n<meta charset=\"utf-8\"/>\n<style>body{background-color:white;}</style>\n\n\n</head>\n<body>\n<div id=\"wkqosdusdm\" style=\"overflow-x:auto;overflow-y:auto;width:auto;height:auto;\">\n  <style>html {\n  font-family: times, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;\n}\n\n#wkqosdusdm .gt_table {\n  display: table;\n  border-collapse: collapse;\n  margin-left: auto;\n  margin-right: auto;\n  color: #333333;\n  font-size: 12px;\n  font-weight: normal;\n  font-style: none;\n  background-color: #FFFFFF;\n  width: auto;\n  border-top-style: none;\n  border-top-width: 2px;\n  border-top-color: #FFFFFF;\n  border-right-style: none;\n  border-right-width: 2px;\n  border-right-color: #D3D3D3;\n  border-bottom-style: none;\n  border-bottom-width: 2px;\n  border-bottom-color: #A8A8A8;\n  border-left-style: none;\n  border-left-width: 2px;\n  border-left-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_heading {\n  background-color: #FFFFFF;\n  text-align: center;\n  border-bottom-color: #FFFFFF;\n  border-left-style: none;\n  border-left-width: 1px;\n  border-left-color: #D3D3D3;\n  border-right-style: none;\n  border-right-width: 1px;\n  border-right-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_title {\n  color: #333333;\n  font-size: 12px;\n  font-weight: initial;\n  padding-top: 4px;\n  padding-bottom: 4px;\n  padding-left: 5px;\n  padding-right: 5px;\n  border-bottom-color: #FFFFFF;\n  border-bottom-width: 0;\n}\n\n#wkqosdusdm .gt_subtitle {\n  color: #333333;\n  font-size: 12px;\n  font-weight: initial;\n  padding-top: 0;\n  padding-bottom: 6px;\n  padding-left: 5px;\n  padding-right: 5px;\n  border-top-color: #FFFFFF;\n  border-top-width: 0;\n}\n\n#wkqosdusdm .gt_bottom_border {\n  border-bottom-style: solid;\n  border-bottom-width: 2px;\n  border-bottom-color: #000000;\n}\n\n#wkqosdusdm .gt_col_headings {\n  border-top-style: none;\n  border-top-width: 2px;\n  border-top-color: #D3D3D3;\n  border-bottom-style: solid;\n  border-bottom-width: 1px;\n  border-bottom-color: #000000;\n  border-left-style: none;\n  border-left-width: 1px;\n  border-left-color: #D3D3D3;\n  border-right-style: none;\n  border-right-width: 1px;\n  border-right-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_col_heading {\n  color: #333333;\n  background-color: #FFFFFF;\n  font-size: 100%;\n  font-weight: normal;\n  text-transform: inherit;\n  border-left-style: none;\n  border-left-width: 1px;\n  border-left-color: #D3D3D3;\n  border-right-style: none;\n  border-right-width: 1px;\n  border-right-color: #D3D3D3;\n  vertical-align: bottom;\n  padding-top: 5px;\n  padding-bottom: 6px;\n  padding-left: 5px;\n  padding-right: 5px;\n  overflow-x: hidden;\n}\n\n#wkqosdusdm .gt_column_spanner_outer {\n  color: #333333;\n  background-color: #FFFFFF;\n  font-size: 100%;\n  font-weight: normal;\n  text-transform: inherit;\n  padding-top: 0;\n  padding-bottom: 0;\n  padding-left: 4px;\n  padding-right: 4px;\n}\n\n#wkqosdusdm .gt_column_spanner_outer:first-child {\n  padding-left: 0;\n}\n\n#wkqosdusdm .gt_column_spanner_outer:last-child {\n  padding-right: 0;\n}\n\n#wkqosdusdm .gt_column_spanner {\n  border-bottom-style: solid;\n  border-bottom-width: 1px;\n  border-bottom-color: #000000;\n  vertical-align: bottom;\n  padding-top: 5px;\n  padding-bottom: 5px;\n  overflow-x: hidden;\n  display: inline-block;\n  width: 100%;\n}\n\n#wkqosdusdm .gt_group_heading {\n  padding-top: 8px;\n  padding-bottom: 8px;\n  padding-left: 5px;\n  padding-right: 5px;\n  color: #333333;\n  background-color: #FFFFFF;\n  font-size: 100%;\n  font-weight: initial;\n  text-transform: inherit;\n  border-top-style: none;\n  border-top-width: 2px;\n  border-top-color: #D3D3D3;\n  border-bottom-style: none;\n  border-bottom-width: 2px;\n  border-bottom-color: #D3D3D3;\n  border-left-style: none;\n  border-left-width: 1px;\n  border-left-color: #D3D3D3;\n  border-right-style: none;\n  border-right-width: 1px;\n  border-right-color: #D3D3D3;\n  vertical-align: middle;\n}\n\n#wkqosdusdm .gt_empty_group_heading {\n  padding: 0.5px;\n  color: #333333;\n  background-color: #FFFFFF;\n  font-size: 100%;\n  font-weight: initial;\n  border-top-style: none;\n  border-top-width: 2px;\n  border-top-color: #D3D3D3;\n  border-bottom-style: none;\n  border-bottom-width: 2px;\n  border-bottom-color: #D3D3D3;\n  vertical-align: middle;\n}\n\n#wkqosdusdm .gt_from_md > :first-child {\n  margin-top: 0;\n}\n\n#wkqosdusdm .gt_from_md > :last-child {\n  margin-bottom: 0;\n}\n\n#wkqosdusdm .gt_row {\n  padding-top: 8px;\n  padding-bottom: 8px;\n  padding-left: 5px;\n  padding-right: 5px;\n  margin: 10px;\n  border-top-style: none;\n  border-top-width: 1px;\n  border-top-color: #FFFFFF;\n  border-left-style: none;\n  border-left-width: 1px;\n  border-left-color: #D3D3D3;\n  border-right-style: none;\n  border-right-width: 1px;\n  border-right-color: #D3D3D3;\n  vertical-align: middle;\n  overflow-x: hidden;\n}\n\n#wkqosdusdm .gt_stub {\n  color: #333333;\n  background-color: #FFFFFF;\n  font-size: 100%;\n  font-weight: initial;\n  text-transform: inherit;\n  border-right-style: none;\n  border-right-width: 2px;\n  border-right-color: #D3D3D3;\n  padding-left: 5px;\n  padding-right: 5px;\n}\n\n#wkqosdusdm .gt_stub_row_group {\n  color: #333333;\n  background-color: #FFFFFF;\n  font-size: 100%;\n  font-weight: initial;\n  text-transform: inherit;\n  border-right-style: none;\n  border-right-width: 2px;\n  border-right-color: #D3D3D3;\n  padding-left: 5px;\n  padding-right: 5px;\n  vertical-align: top;\n}\n\n#wkqosdusdm .gt_row_group_first td {\n  border-top-width: 2px;\n}\n\n#wkqosdusdm .gt_summary_row {\n  color: #333333;\n  background-color: #FFFFFF;\n  text-transform: inherit;\n  padding-top: 8px;\n  padding-bottom: 8px;\n  padding-left: 5px;\n  padding-right: 5px;\n}\n\n#wkqosdusdm .gt_first_summary_row {\n  border-top-style: none;\n  border-top-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_first_summary_row.thick {\n  border-top-width: 2px;\n}\n\n#wkqosdusdm .gt_last_summary_row {\n  padding-top: 8px;\n  padding-bottom: 8px;\n  padding-left: 5px;\n  padding-right: 5px;\n  border-bottom-style: none;\n  border-bottom-width: 2px;\n  border-bottom-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_grand_summary_row {\n  color: #333333;\n  background-color: #FFFFFF;\n  text-transform: inherit;\n  padding-top: 8px;\n  padding-bottom: 8px;\n  padding-left: 5px;\n  padding-right: 5px;\n}\n\n#wkqosdusdm .gt_first_grand_summary_row {\n  padding-top: 8px;\n  padding-bottom: 8px;\n  padding-left: 5px;\n  padding-right: 5px;\n  border-top-style: none;\n  border-top-width: 6px;\n  border-top-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_striped {\n  background-color: rgba(128, 128, 128, 0.05);\n}\n\n#wkqosdusdm .gt_table_body {\n  border-top-style: solid;\n  border-top-width: 1px;\n  border-top-color: #000000;\n  border-bottom-style: solid;\n  border-bottom-width: 1px;\n  border-bottom-color: #000000;\n}\n\n#wkqosdusdm .gt_footnotes {\n  color: #333333;\n  background-color: #FFFFFF;\n  border-bottom-style: none;\n  border-bottom-width: 2px;\n  border-bottom-color: #D3D3D3;\n  border-left-style: none;\n  border-left-width: 2px;\n  border-left-color: #D3D3D3;\n  border-right-style: none;\n  border-right-width: 2px;\n  border-right-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_footnote {\n  margin: 0px;\n  font-size: 90%;\n  padding-left: 4px;\n  padding-right: 4px;\n  padding-left: 5px;\n  padding-right: 5px;\n}\n\n#wkqosdusdm .gt_sourcenotes {\n  color: #333333;\n  background-color: #FFFFFF;\n  border-bottom-style: none;\n  border-bottom-width: 2px;\n  border-bottom-color: #D3D3D3;\n  border-left-style: none;\n  border-left-width: 2px;\n  border-left-color: #D3D3D3;\n  border-right-style: none;\n  border-right-width: 2px;\n  border-right-color: #D3D3D3;\n}\n\n#wkqosdusdm .gt_sourcenote {\n  font-size: 90%;\n  padding-top: 4px;\n  padding-bottom: 4px;\n  padding-left: 5px;\n  padding-right: 5px;\n}\n\n#wkqosdusdm .gt_left {\n  text-align: left;\n}\n\n#wkqosdusdm .gt_center {\n  text-align: center;\n}\n\n#wkqosdusdm .gt_right {\n  text-align: right;\n  font-variant-numeric: tabular-nums;\n}\n\n#wkqosdusdm .gt_font_normal {\n  font-weight: normal;\n}\n\n#wkqosdusdm .gt_font_bold {\n  font-weight: bold;\n}\n\n#wkqosdusdm .gt_font_italic {\n  font-style: italic;\n}\n\n#wkqosdusdm .gt_super {\n  font-size: 65%;\n}\n\n#wkqosdusdm .gt_footnote_marks {\n  font-style: italic;\n  font-weight: normal;\n  font-size: 75%;\n  vertical-align: 0.4em;\n}\n\n#wkqosdusdm .gt_asterisk {\n  font-size: 100%;\n  vertical-align: 0;\n}\n\n#wkqosdusdm .gt_slash_mark {\n  font-size: 0.7em;\n  line-height: 0.7em;\n  vertical-align: 0.15em;\n}\n\n#wkqosdusdm .gt_fraction_numerator {\n  font-size: 0.6em;\n  line-height: 0.6em;\n  vertical-align: 0.45em;\n}\n\n#wkqosdusdm .gt_fraction_denominator {\n  font-size: 0.6em;\n  line-height: 0.6em;\n  vertical-align: -0.05em;\n}\n</style>\n  <table class=\"gt_table\">\n  \n  <thead class=\"gt_col_headings\">\n    <tr>\n      <th class=\"gt_col_heading gt_columns_bottom_border gt_left\" rowspan=\"1\" colspan=\"1\"></th>\n      <th class=\"gt_col_heading gt_columns_bottom_border gt_right\" rowspan=\"1\" colspan=\"1\"><em><center>OR [95% CI]</center></em></th>\n      <th class=\"gt_col_heading gt_columns_bottom_border gt_right\" rowspan=\"1\" colspan=\"1\"><em><center>Stand. OR [95% CI]</center></em></th>\n    </tr>\n  </thead>\n  <tbody class=\"gt_table_body\">\n    <tr><td class=\"gt_row gt_left\"><div class='gt_from_md'><p>religionyes</p>\n</div></td>\n<td class=\"gt_row gt_right\"><div class='gt_from_md'><p>0.95 &nbsp;&nbsp;&nbsp; [0.82, 1.09]</p>\n</div></td>\n<td class=\"gt_row gt_right\"><div class='gt_from_md'><p>0.95 &nbsp;&nbsp;&nbsp; [0.82, 1.09]</p>\n</div></td></tr>\n    <tr><td class=\"gt_row gt_left\"><div class='gt_from_md'><p>age</p>\n</div></td>\n<td class=\"gt_row gt_right\"><div class='gt_from_md'><p>1.01 *** [1.01, 1.02]</p>\n</div></td>\n<td class=\"gt_row gt_right\"><div class='gt_from_md'><p>1.25 *** [1.19, 1.32]</p>\n</div></td></tr>\n    <tr><td class=\"gt_row gt_left\"><div class='gt_from_md'><p>gendermale</p>\n</div></td>\n<td class=\"gt_row gt_right\"><div class='gt_from_md'><p>1.16 **&nbsp; [1.05, 1.29]</p>\n</div></td>\n<td class=\"gt_row gt_right\"><div class='gt_from_md'><p>1.16 **&nbsp; [1.05, 1.29]</p>\n</div></td></tr>\n   <tr style=\"border-top-style: solid; border-top-width: 2px;\">\n    <td class=\"gt_row gt_left\" rowspan=\"1\" colspan=\"1\">  <em>N     </td> <td class=\"gt_row gt_center\" rowspan=\"1\" colspan=\"2\"> 5381   </td> </tr> <tr>\n    <td class=\"gt_row gt_left\" rowspan=\"1\" colspan=\"1\">  <em>R<sup>2</sup>     </td> <td class=\"gt_row gt_center\" rowspan=\"1\" colspan=\"2\"> .02   </td> </tr> </tbody>\n  <tfoot class=\"gt_sourcenotes\">\n    <tr>\n      <td class=\"gt_sourcenote\" colspan=\"3\"></td>\n    </tr>\n    <tr>\n      <td class=\"gt_sourcenote\" colspan=\"3\">Given that dummy variables lose their interpretability when standardized, only continuous predictors are standardized.</td>\n    </tr>\n    <tr>\n      <td class=\"gt_sourcenote\" colspan=\"3\">&dagger; <em>p</em> &lt; .1, * <em>p</em> &lt; .05, ** <em>p</em> &lt; .01, *** <em>p</em> &lt; .001</td>\n    </tr>\n  </tfoot>\n  \n</table>\n</div>\n</body>\n</html>\n"
#> 
#> attr(,"class")
#> [1] "timesaveR_raw_html" "list"