Creates a summary table comparing standardised and non-standardised linear models

This function creates a summary table for lm models (including mice::mira objects containing lm-models) that shows a standardised and non-standardised version of the model side-by-side. Several pairs of such models can be compared side-by-side.

report_lm_with_std(
  mod,
  mod_std,
  conf_level = 0.95,
  coef_renames = NULL,
  fmt = "%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  R2_change = FALSE,
  notes = list(NULL),
  apa_style = TRUE,
  stars = std_stars_pad,
  ...
)

Arguments

mod
A lm-model/mira object of lm models, with variables not standardised (or a list of such models)

mod_std
A lm-model/mira object of lm models, with standardised variables. Can be created with lm_std (or a list of such models)

conf_level
Confidence level to use for confidence intervals, defaults to .95

coef_renames
A named character vector with new names for the coefficients or a tibble as provided by get_coef_rename_tribble for variables. If NULL, then the coefficients are not renamed.

fmt
determines how to format numeric values

integer: the number of digits to keep after the period `format(round(x, fmt), nsmall=fmt)`

character: passed to the `sprintf` function (e.g., '%.3f' keeps 3 digits with trailing zero). See `?sprintf`

function: returns a formatted character string.

statistic_vertical
Should standard errors and CIs be shown below coefficients? Defaults to horizontal layout

filename
the file name to create a HTML file on disk.

model_names
If several pairs of models are to be plotted side by side, indicate the label for each *pair* here

show_nimp
Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic

R2_change
Logical. Report R2 change and F-test to compare models. Only implemented for comparing two pairs of models.

notes
List of notes to append to bottom of table. An explanation of significance stars is automatically added. If the std models were run with a helper function in this package, a note regarding the standardisation is also automatically added.

apa_style
Logical, should APA-style formatting be applied

stars
Named vector of significance stars and their thresholds, check `rNuggets:::std_stars_pad` for default.

...
Arguments passed on to modelsummary::modelsummary

coef_map

named character vector. Values refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term. Coefficients that are omitted from this vector will be omitted from the table. The table will be ordered in the same order as this vector.

coef_omit

string regular expression. Omits all matching coefficients from the table using `grepl(perl=TRUE)`.

coef_rename

named character vector. Values refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term.

Value

A list with `gt_tab` (the gt-table object including the parts of the table that can be created with gt. This can be post-processed and formatted with functions in the gt-package, but does not include the lower part with model statistics, e.g., R^2.) and `html_code` (the code that creates the full table, and is used to render it in the Viewer).

Examples

# Standard lm model
mod1 <- lm(mpg ~ hp + wt, mtcars)

# Model with standardised coefficients

mod2 <- lm_std(mpg ~ hp + wt, mtcars)

report_lm_with_std(mod1, mod2)

#> <!DOCTYPE html>
#> <html lang="en">
#> <head>
#> <meta charset="utf-8"/>
#> <style>body{background-color:white;}</style>
#> 
#> 
#> </head>
#> <body>
#> <style>html {
#>   font-family: times, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #ulbfdpbcyv .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: normal;
#>   font-style: none;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: white;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #ulbfdpbcyv .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ulbfdpbcyv .gt_title {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ulbfdpbcyv .gt_subtitle {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ulbfdpbcyv .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: black;
#> }
#> 
#> #ulbfdpbcyv .gt_col_headings {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: black;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ulbfdpbcyv .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #ulbfdpbcyv .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #ulbfdpbcyv .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #ulbfdpbcyv .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #ulbfdpbcyv .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: black;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #ulbfdpbcyv .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ulbfdpbcyv .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ulbfdpbcyv .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #ulbfdpbcyv .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #ulbfdpbcyv .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: none;
#>   border-top-width: 1px;
#>   border-top-color: white;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #ulbfdpbcyv .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #ulbfdpbcyv .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ulbfdpbcyv .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ulbfdpbcyv .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ulbfdpbcyv .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: none;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ulbfdpbcyv .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #ulbfdpbcyv .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: black;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: black;
#> }
#> 
#> #ulbfdpbcyv .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ulbfdpbcyv .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #ulbfdpbcyv .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ulbfdpbcyv .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #ulbfdpbcyv .gt_left {
#>   text-align: left;
#> }
#> 
#> #ulbfdpbcyv .gt_center {
#>   text-align: center;
#> }
#> 
#> #ulbfdpbcyv .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ulbfdpbcyv .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ulbfdpbcyv .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ulbfdpbcyv .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ulbfdpbcyv .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ulbfdpbcyv .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="ulbfdpbcyv" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1"><em><center>B (SE)</center></em></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1"><em><center>&beta; [95% CI]</center></em></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr>
#>       <td class="gt_row gt_left"><div class='gt_from_md'><p>(Intercept)</p>
#> </div></td>
#>       <td class="gt_row gt_right"><div class='gt_from_md'><p>37.23 (1.60)***</p>
#> </div></td>
#>       <td class="gt_row gt_right"><div class='gt_from_md'><p>0.00 [-0.16, 0.16]</p>
#> </div></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left"><div class='gt_from_md'><p>hp</p>
#> </div></td>
#>       <td class="gt_row gt_right"><div class='gt_from_md'><p>-0.03 (0.01)**&nbsp;</p>
#> </div></td>
#>       <td class="gt_row gt_right"><div class='gt_from_md'><p>-0.36 [-0.57, -0.15]</p>
#> </div></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left"><div class='gt_from_md'><p>wt</p>
#> </div></td>
#>       <td class="gt_row gt_right"><div class='gt_from_md'><p>-3.88 (0.63)***</p>
#> </div></td>
#>       <td class="gt_row gt_right"><div class='gt_from_md'><p>-0.63 [-0.84, -0.42]</p>
#> </div></td>
#>     </tr>
#>    <tr style="border-top-style: solid; border-top-width: 2px;">
#>     <td class="gt_row gt_left" rowspan="1" colspan="1">  <em>N     </td> <td class="gt_row gt_center" rowspan="1" colspan="2"> 32   </td> </tr> <tr>
#>     <td class="gt_row gt_left" rowspan="1" colspan="1">  <em>R<sup>2</sup>     </td> <td class="gt_row gt_center" rowspan="1" colspan="2"> .83   </td> </tr> <tr>
#>     <td class="gt_row gt_left" rowspan="1" colspan="1"><em>F</em>-tests</td> <td class="gt_row gt_center" rowspan="1" colspan="2"> <div class='gt_from_md'><p><em>F</em>(2, 29) = 69.21, <em>p</em> &lt; .001</p>
#> </div>   </td> </tr> </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3">Given that dummy variables lose their interpretability when standardised (Fox, 2015), &beta; for dummy variables are semi-standardised, indicating the impact of that dummy on the standardised outcome variable.</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3">&dagger; <em>p</em> &lt; .1, * <em>p</em> &lt; .05, ** <em>p</em> &lt; .01, *** <em>p</em> &lt; .001</td>
#>     </tr>
#>   </tfoot>
#>   
#> </table></div>
#> </body>
#> </html>
#>