Create a summary table comparing standardized and non-standardized proportional odd logistic regression models

This function creates a summary table for polr models (including mice::mira objects containing polr-models) that shows a standardized and non-standardized version of the model side-by-side. Several pairs of such models can be compared side-by-side.

report_polr_with_std(
  mod,
  mod_std,
  OR = TRUE,
  conf_level = 0.95,
  fmt = "%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  notes = list(),
  apa_style = TRUE,
  stars = std_stars_pad,
  ...
)

Arguments

mod
A polr-model/mira object of polr models, with variables not standardized (or a list of such models)

mod_std
A polr-model/mira object of polr models, with standardized predictor variables (or a list of such models)

OR
Logical. Should odds ratios be shown instead of typical coefficients. If TRUE, estimates are exponentiated

conf_level
Confidence level to use for confidence intervals, defaults to .95

fmt
determines how to format numeric values

integer: the number of digits to keep after the period format(round(x, fmt), nsmall=fmt)

character: passed to the sprintf function (e.g., '\

function: returns a formatted character string.

statistic_vertical
Should standard errors and CIs be shown below coefficients? Defaults to horizontal layout

filename
the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)

model_names
If several pairs of models are to be plotted side by side, indicate the label for each *pair* here

show_nimp
Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic

notes
List of notes to append to bottom of table. An explanation of significance stars is automatically added. A note is also added stating that dummy variables were not scaled in standardization. If you approached standardisation differently, that should be removed.

apa_style
Logical, should APA-style formatting be applied

stars
Named vector of significance stars and their thresholds, check `timesaveR:::std_stars_pad` for default.

...
Arguments passed on to modelsummary::modelsummary

output

filename or object type (character string)

Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "modelsummary_list", "gt", "kableExtra", "huxtable", "flextable".

To change the default output format, type options(modelsummary_default = "latex"), where latex can be any of the valid object types listed above.

Warning: the output argument cannot be used when customizing tables with external packages.

See the 'Details' section below for more information.

estimate

string or glue string of the estimate to display (or a vector with one string per model). Valid entries include any column name of the data.frame produced by get_estimates(model). Examples:

"estimate"

"{estimate} ({std.error}){stars}"

"{estimate} [{conf.low}, {conf.high}]"

vcov

robust standard errors and other manual statistics. The vcov argument accepts five types of input (see the 'Details' and 'Examples' sections below):

string, vector, or list of strings: "robust", "HC", "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5", "stata", or "classical" (alias "constant" or "iid").

formula or list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).

function or list of functions which return variance-covariance matrices with row and column names equal to the names of your coefficient estimates (e.g., stats::vcov, sandwich::vcovHC).

list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below. Warning: since this list of vectors can include arbitrary strings or numbers, modelsummary cannot automatically calculate p values. The stars argument may thus use incorrect significance thresholds when vcov is a list of vectors.

coef_map

character vector. Subset, rename, and reorder coefficients. Coefficients omitted from this vector are omitted from the table. The order of the vector determines the order of the table. coef_map can be a named or an unnamed character vector (see the Examples section below). If coef_map is a named vector, its values define the labels that must appear in the table, and its names identify the original term names stored in the model object: c("hp:mpg"="HPxM/G").

coef_omit

string regular expression. Omits all matching coefficients from the table using grepl(perl=TRUE).

coef_rename

named character vector or function which returns a named vector. Values of the vector refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term.

gof_map

NULL (default): the modelsummary::gof_map dictionary is used for formatting, and all unknown statistic are included.

data.frame with 3 columns named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples' section below.

list of lists, each of which includes 3 elements named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples section below'.

gof_omit

string regular expression. Omits all matching gof statistics from the table (using grepl(perl=TRUE)).

group

a two-sided formula with two or three components which describes how groups of parameters should be displayed. The formula must include both a "term" and a "model" component. In addition, a component can be used to identify groups of parameters (e.g., outcome levels of a multinomial logit model). This group identifier must be the name of a column in the data.frame produced by get_estimates(model).

term ~ model displays coefficients as rows and models as columns

model ~ term displays models as rows and coefficients as columns

response + term ~ model displays response levels and coefficients as rows and models as columns.

group_map

named or unnamed character vector. Subset, rename, and reorder coefficient groups specified in the group argument. See coef_map.

add_rows

a data.frame (or tibble) with the same number of columns as your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the row positions. See Examples section below.

align

A character string of length equal to the number of columns in the table. "lcr" means that the first column will be left-aligned, the 2nd column center-aligned, and the 3rd column right-aligned.

title

string

Details

The R2 shown is the maximum likelihood pseudo R2 returned by pR2.

Examples

library(MASS)
pov_att <- polr(poverty ~ religion + age + gender, data = WVS)
pov_att_std <- polr_std(poverty ~ religion + age + gender, data = WVS)
report_polr_with_std(pov_att, pov_att_std, coef_omit = "\\|")

#> 
#> Re-fitting to get Hessian

#> fitting null model for pseudo-r2

#> <!DOCTYPE html>
#> <html lang="en">
#> <head>
#> <meta charset="utf-8"/>
#> <style>body{background-color:white;}</style>
#> 
#> 
#> </head>
#> <body>
#> <div id="vuzjcdmfnb" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: times, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #vuzjcdmfnb .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: normal;
#>   font-style: none;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: white;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #vuzjcdmfnb .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vuzjcdmfnb .gt_title {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #vuzjcdmfnb .gt_subtitle {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #vuzjcdmfnb .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: black;
#> }
#> 
#> #vuzjcdmfnb .gt_col_headings {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: black;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vuzjcdmfnb .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #vuzjcdmfnb .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #vuzjcdmfnb .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #vuzjcdmfnb .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #vuzjcdmfnb .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: black;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #vuzjcdmfnb .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #vuzjcdmfnb .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #vuzjcdmfnb .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #vuzjcdmfnb .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #vuzjcdmfnb .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: none;
#>   border-top-width: 1px;
#>   border-top-color: white;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #vuzjcdmfnb .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #vuzjcdmfnb .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #vuzjcdmfnb .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #vuzjcdmfnb .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #vuzjcdmfnb .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: none;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #vuzjcdmfnb .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #vuzjcdmfnb .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: black;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: black;
#> }
#> 
#> #vuzjcdmfnb .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vuzjcdmfnb .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #vuzjcdmfnb .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #vuzjcdmfnb .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #vuzjcdmfnb .gt_left {
#>   text-align: left;
#> }
#> 
#> #vuzjcdmfnb .gt_center {
#>   text-align: center;
#> }
#> 
#> #vuzjcdmfnb .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #vuzjcdmfnb .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #vuzjcdmfnb .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #vuzjcdmfnb .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #vuzjcdmfnb .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #vuzjcdmfnb .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 65%;
#> }
#> </style>
#>   <table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1"></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1"><em><center>OR [95% CI]</center></em></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1"><em><center>Stand. OR [95% CI]</center></em></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_left"><div class='gt_from_md'><p>religionyes</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>0.95 &nbsp;&nbsp;&nbsp; [0.82, 1.09]</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>0.95 &nbsp;&nbsp;&nbsp; [0.82, 1.09]</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_left"><div class='gt_from_md'><p>age</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.01 *** [1.01, 1.02]</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.25 *** [1.19, 1.32]</p>
#> </div></td></tr>
#>     <tr><td class="gt_row gt_left"><div class='gt_from_md'><p>gendermale</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.16 **&nbsp; [1.05, 1.29]</p>
#> </div></td>
#> <td class="gt_row gt_right"><div class='gt_from_md'><p>1.16 **&nbsp; [1.05, 1.29]</p>
#> </div></td></tr>
#>    <tr style="border-top-style: solid; border-top-width: 2px;">
#>     <td class="gt_row gt_left" rowspan="1" colspan="1">  <em>N     </td> <td class="gt_row gt_center" rowspan="1" colspan="2"> 5381   </td> </tr> <tr>
#>     <td class="gt_row gt_left" rowspan="1" colspan="1">  <em>R<sup>2</sup>     </td> <td class="gt_row gt_center" rowspan="1" colspan="2"> .02   </td> </tr> </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3">Given that dummy variables lose their interpretability when standardized, only continuous predictors are standardized.</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3">&dagger; <em>p</em> &lt; .1, * <em>p</em> &lt; .05, ** <em>p</em> &lt; .01, *** <em>p</em> &lt; .001</td>
#>     </tr>
#>   </tfoot>
#>   
#> </table>
#> </div>
#> </body>
#> </html>
#>