Create a summary table comparing standardized and non-standardized proportional odd logistic regression models

This function creates a summary table for polr models (including mice::mira objects containing polr-models) that shows a standardized and non-standardized version of the model side-by-side. Several pairs of such models can be compared side-by-side.

report_polr_with_std(
  mod,
  mod_std,
  OR = TRUE,
  conf_level = 0.95,
  fmt = "%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  notes = list(),
  apa_style = TRUE,
  stars = std_stars_pad,
  ...
)

Arguments

mod

A polr-model/mira object of polr models, with variables not standardized (or a list of such models)

mod_std

A polr-model/mira object of polr models, with standardized predictor variables (or a list of such models)

OR

Logical. Should odds ratios be shown instead of typical coefficients. If TRUE, estimates are exponentiated

conf_level

Confidence level to use for confidence intervals, defaults to .95

fmt

how to format numeric values: integer, user-supplied function, or modelsummary function.

Integer: Number of decimal digits

User-supplied functions:

Any function which accepts a numeric vector and returns a character vector of the same length.

modelsummary functions:

fmt = fmt_significant(2): Two significant digits (at the term-level)

fmt = fmt_decimal(digits = 2, pdigits = 3): Decimal digits for estimate and p values

fmt = fmt_sprintf("%.3f"): See ?sprintf

fmt = fmt_term("(Intercept)" = 1, "X" = 2): Format terms differently

fmt = fmt_statistic("estimate" = 1, "r.squared" = 6): Format statistics differently.

fmt = fmt_identity(): unformatted raw values

string:

Note on LaTeX output: To ensure proper typography, all numeric entries are enclosed in the \num{} command, which requires the siunitx package to be loaded in the LaTeX preamble. This behavior can be altered with global options. See the 'Details' section.

statistic_vertical

Should standard errors and CIs be shown below coefficients? Defaults to horizontal layout

filename

the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)

model_names

If several pairs of models are to be plotted side by side, indicate the label for each pair here

show_nimp

Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic

notes

List of notes to append to bottom of table. An explanation of significance stars is automatically added. A note is also added stating that dummy variables were not scaled in standardization. If you approached standardisation differently, that should be removed.

apa_style

Logical, should APA-style formatting be applied

stars

Named vector of significance stars and their thresholds, check timesaveR:::std_stars_pad for default.

...

Arguments passed on to modelsummary::modelsummary

output

filename or object type (character string)

Supported filename extensions: .docx, .html, .tex, .md, .txt, .csv, .xlsx, .png, .jpg

Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "typst", "data.frame", "tinytable", "gt", "kableExtra", "huxtable", "flextable", "DT", "jupyter". The "modelsummary_list" value produces a lightweight object which can be saved and fed back to the modelsummary function.

The "default" output format can be set to "tinytable", "kableExtra", "gt", "flextable", "huxtable", "DT", or "markdown"

If the user does not choose a default value, the packages listed above are tried in sequence.

Session-specific configuration: options("modelsummary_factory_default" = "gt")

Persistent configuration: config_modelsummary(output = "markdown")

Warning: Users should not supply a file name to the output argument if they intend to customize the table with external packages. See the 'Details' section.

LaTeX compilation requires the booktabs and siunitx packages, but siunitx can be disabled or replaced with global options. See the 'Details' section.

estimate

a single string or a character vector of length equal to the number of models. Valid entries include any column name of the data.frame produced by get_estimates(model), and strings with curly braces compatible with the glue package format. Examples:

"estimate"

"{estimate} ({std.error}){stars}"

"{estimate} [{conf.low}, {conf.high}]"

vcov

robust standard errors and other manual statistics. The vcov argument accepts six types of input (see the 'Details' and 'Examples' sections below):

NULL returns the default uncertainty estimates of the model object

string, vector, or (named) list of strings. "iid", "classical", and "constant" are aliases for NULL, which returns the model's default uncertainty estimates. The strings "HC", "HC0", "HC1" (alias: "stata"), "HC2", "HC3" (alias: "robust"), "HC4", "HC4m", "HC5", "HAC", "NeweyWest", "Andrews", "panel-corrected", "outer-product", and "weave" use variance-covariance matrices computed using functions from the sandwich package, or equivalent method. "BS", "bootstrap", "residual", "mammen", "webb", "xy", "wild" use the sandwich::vcovBS(). The behavior of those functions can (and sometimes must) be altered by passing arguments to sandwich directly from modelsummary through the ellipsis (...), but it is safer to define your own custom functions as described in the next bullet.

function or (named) list of functions which return variance-covariance matrices with row and column names equal to the names of your coefficient estimates (e.g., stats::vcov, sandwich::vcovHC, function(x) vcovPC(x, cluster="country")).

formula or (named) list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).

named list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a named list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below. Warning: since this list of vectors can include arbitrary strings or numbers, modelsummary cannot automatically calculate p values. The stars argument may thus use incorrect significance thresholds when vcov is a list of vectors.

exponentiate

TRUE, FALSE, or logical vector of length equal to the number of models. If TRUE, the estimate, conf.low, and conf.high statistics are exponentiated, and the std.error is transformed to exp(estimate)*std.error.

shape

NULL, formula, or string which determines the shape of a table.

NULL: Default shape with terms in rows and models in columns.

Formula: The left side determines what appears on rows, and the right side determines what appears on columns. The formula can include one or more group identifier(s) to display related terms together, which can be useful for models with multivariate outcomes or grouped coefficients (See examples section below). The group identifier(s) must be column names produced by: get_estimates(model). The group identifier(s) can be combined with the term identifier in a single column by using the colon to represent an interaction. If an incomplete formula is supplied (e.g., ~statistic), modelsummary tries to complete it automatically. Goodness-of-fit statistics are only appended to the bottom of the table when model is on the right hand side of the formula (i.e., columns). Potential shape values include:

term + statistic ~ model: default

term ~ model + statistic: statistics in separate columns

model + statistic ~ term: models in rows and terms in columns

term + response + statistic ~ model: term and group id in separate columns

term : response + statistic ~ model: term and group id in a single column

term ~ response

String: "cbind", "rbind", "rcollapse"

"cbind": side-by-side models with autmoatic spanning column headers to group models (tinytable only feature).

"rbind" or "rcollapse": "panels" or "stacks" of regression models.

the models argument must be a (potentially named) nested list of models.

Unnamed nested list with 2 panels: list(list(model1, model2), list(model3, model4))

Named nested list with 2 panels: list("Panel A" = list(model1, model2), "Panel B" = list(model3, model4))

Named panels and named models: list("Panel A" = list("(I)" = model1, "(II)" = model2), "Panel B" = list("(I)" = model3, "(II)" = model4))

"rbind": Bind the rows of independent regression tables

"rcollapse": Bind the rows of regression tables and create a panel at the bottom where we "collapse" goodness-of-fit statistics which are identical across models.

coef_map

character vector. Subset, rename, and reorder coefficients. Coefficients omitted from this vector are omitted from the table. The order of the vector determines the order of the table. coef_map can be a named or an unnamed character vector. If coef_map is a named vector, its values define the labels that must appear in the table, and its names identify the original term names stored in the model object: c("hp:mpg"="HPxM/G"). See Examples section below.

coef_omit

integer vector or regular expression to identify which coefficients to omit (or keep) from the table. Positive integers determine which coefficients to omit. Negative integers determine which coefficients to keep. A regular expression can be used to omit coefficients, and perl-compatible "negative lookaheads" can be used to specify which coefficients to keep in the table. Examples:

c(2, 3, 5): omits the second, third, and fifth coefficients.

c(-2, -3, -5): negative values keep the second, third, and fifth coefficients.

"ei": omit coefficients matching the "ei" substring.

"^Volume$": omit the "Volume" coefficient.

"ei|rc": omit coefficients matching either the "ei" or the "rc" substrings.

"^(?!Vol)": keep coefficients starting with "Vol" (inverse match using a negative lookahead).

"^(?!.*ei)": keep coefficients matching the "ei" substring.

"^(?!.*ei|.*pt)": keep coefficients matching either the "ei" or the "pt" substrings.

See the Examples section below for complete code.

coef_rename

logical, named or unnamed character vector, or function

Logical: TRUE renames variables based on the "label" attribute of each column. See the Example section below.

Unnamed character vector of length equal to the number of coefficients in the final table, after coef_omit is applied.

Named character vector: Values refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object. Ex: c("hp:mpg"="hp X mpg")

Function: Accepts a character vector of the model's term names and returns a named vector like the one described above. The modelsummary package supplies a coef_rename() function which can do common cleaning tasks: modelsummary(model, coef_rename = coef_rename)

gof_map

rename, reorder, and omit goodness-of-fit statistics and other model information. This argument accepts 4 types of values:

NULL (default): the modelsummary::gof_map dictionary is used for formatting, and all unknown statistic are included.

character vector: "all", "none", or a vector of statistics such as c("rmse", "nobs", "r.squared"). Elements correspond to colnames in the data.frame produced by get_gof(model). The modelsummary::gof_map default dictionary is used to format and rename statistics.

NA: excludes all statistics from the bottom part of the table.

data.frame with 3 columns named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples' section below. The fmt column in this data frame only accepts integers. For more flexibility, use a list of lists, as described in the next bullet.

list of lists, each of which includes 3 elements named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples section below'.

gof_omit

string regular expression (perl-compatible) used to determine which statistics to omit from the bottom section of the table. A "negative lookahead" can be used to specify which statistics to keep in the table. Examples:

"IC": omit statistics matching the "IC" substring.

"BIC|AIC": omit statistics matching the "AIC" or "BIC" substrings.

"^(?!.*IC)": keep statistics matching the "IC" substring.

gof_function

function which accepts a model object in the model argument and returns a 1-row data.frame with one custom goodness-of-fit statistic per column.

group_map

named or unnamed character vector. Subset, rename, and reorder coefficient groups specified a grouping variable specified in the shape argument formula. This argument behaves like coef_map.

add_columns

a data.frame (or tibble) with the same number of rows as #' your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the columns positions. See Examples section below.

add_rows

a data.frame (or tibble) with the same number of columns as your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the row positions. See Examples section below.

align

A string with a number of characters equal to the number of columns in the table (e.g., align = "lcc"). Valid characters: l, c, r, d.

"l": left-aligned column

"c": centered column

"r": right-aligned column

"d": dot-aligned column. For LaTeX/PDF output, this option requires at least version 3.0.25 of the siunitx LaTeX package. See the LaTeX preamble help section below for commands to insert in your LaTeX preamble.

title

string. Cross-reference labels should be added with Quarto or Rmarkdown chunk options when applicable. When saving standalone LaTeX files, users can add a label such as \\label{tab:mytable} directly to the title string, while also specifying escape=FALSE.

escape

boolean TRUE escapes or substitutes LaTeX/HTML characters which could prevent the file from compiling/displaying. TRUE escapes all cells, captions, and notes. Users can have more fine-grained control by setting escape=FALSE and using an external command such as: modelsummary(model, "latex") |> tinytable::format_tt(tab, j=1:5, escape=TRUE)

Details

The R2 shown is the maximum likelihood pseudo R2 returned by pR2.

Examples

library(MASS)
#> 
#> Attaching package: ‘MASS’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
pov_att <- polr(poverty ~ religion + age + gender, data = WVS)
pov_att_std <- polr_std(poverty ~ religion + age + gender, data = WVS)
report_polr_with_std(pov_att, pov_att_std, coef_omit = "\\|")
#> fitting null model for pseudo-r2
#> <!DOCTYPE html>
#> <html lang="en">
#> <head>
#> <meta charset="utf-8"/>
#> <style>body{background-color:white;}</style>
#> 
#> 
#> </head>
#> <body>
#> <div id="ojkdxuukad" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#ojkdxuukad table {
#>   font-family: times, system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #ojkdxuukad thead, #ojkdxuukad tbody, #ojkdxuukad tfoot, #ojkdxuukad tr, #ojkdxuukad td, #ojkdxuukad th {
#>   border-style: none;
#> }
#> 
#> #ojkdxuukad p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #ojkdxuukad .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: normal;
#>   font-style: none;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #FFFFFF;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #ojkdxuukad .gt_title {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ojkdxuukad .gt_subtitle {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ojkdxuukad .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #ojkdxuukad .gt_col_headings {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #000000;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #ojkdxuukad .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #ojkdxuukad .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #ojkdxuukad .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #ojkdxuukad .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #000000;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #ojkdxuukad .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #ojkdxuukad .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #ojkdxuukad .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ojkdxuukad .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #ojkdxuukad .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #ojkdxuukad .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: none;
#>   border-top-width: 1px;
#>   border-top-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #ojkdxuukad .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ojkdxuukad .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #ojkdxuukad .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #ojkdxuukad .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #ojkdxuukad .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ojkdxuukad .gt_first_summary_row {
#>   border-top-style: none;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #ojkdxuukad .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ojkdxuukad .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: none;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: none;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #ojkdxuukad .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #000000;
#>   border-bottom-style: solid;
#>   border-bottom-width: 1px;
#>   border-bottom-color: #000000;
#> }
#> 
#> #ojkdxuukad .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ojkdxuukad .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ojkdxuukad .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ojkdxuukad .gt_left {
#>   text-align: left;
#> }
#> 
#> #ojkdxuukad .gt_center {
#>   text-align: center;
#> }
#> 
#> #ojkdxuukad .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ojkdxuukad .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ojkdxuukad .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ojkdxuukad .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ojkdxuukad .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ojkdxuukad .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #ojkdxuukad .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #ojkdxuukad .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #ojkdxuukad .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #ojkdxuukad .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #ojkdxuukad .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #ojkdxuukad .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> 
#> #ojkdxuukad .katex-display {
#>   display: inline-flex !important;
#>   margin-bottom: 0.75em !important;
#> }
#> 
#> #ojkdxuukad div.Reactable > div.rt-table > div.rt-thead > div.rt-tr.rt-tr-group-header > div.rt-th-group:after {
#>   height: 0px !important;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="&lt;span class='gt_from_md'&gt;&lt;/span&gt;"><span class='gt_from_md'></span></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="&lt;span class='gt_from_md'&gt;&lt;em&gt;&lt;center&gt;OR [95% CI]&lt;/center&gt;&lt;/em&gt;&lt;/span&gt;"><span class='gt_from_md'><em><center>OR [95% CI]</center></em></span></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="&lt;span class='gt_from_md'&gt;&lt;em&gt;&lt;center&gt;Stand. OR [95% CI]&lt;/center&gt;&lt;/em&gt;&lt;/span&gt;"><span class='gt_from_md'><em><center>Stand. OR [95% CI]</center></em></span></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers=" " class="gt_row gt_left"><span class='gt_from_md'>religionyes</span></td>
#> <td headers="Model1" class="gt_row gt_right"><span class='gt_from_md'>0.95 &nbsp;&nbsp;&nbsp; [0.82, 1.09]</span></td>
#> <td headers="Model2" class="gt_row gt_right"><span class='gt_from_md'>0.95 &nbsp;&nbsp;&nbsp; [0.82, 1.09]</span></td></tr>
#>     <tr><td headers=" " class="gt_row gt_left"><span class='gt_from_md'>age</span></td>
#> <td headers="Model1" class="gt_row gt_right"><span class='gt_from_md'>1.01 *** [1.01, 1.02]</span></td>
#> <td headers="Model2" class="gt_row gt_right"><span class='gt_from_md'>1.25 *** [1.19, 1.32]</span></td></tr>
#>     <tr><td headers=" " class="gt_row gt_left"><span class='gt_from_md'>gendermale</span></td>
#> <td headers="Model1" class="gt_row gt_right"><span class='gt_from_md'>1.16 **&nbsp; [1.05, 1.29]</span></td>
#> <td headers="Model2" class="gt_row gt_right"><span class='gt_from_md'>1.16 **&nbsp; [1.05, 1.29]</span></td></tr>
#>    <tr style="border-top-style: solid; border-top-width: 2px;">
#>     <td class="gt_row gt_left" rowspan="1" colspan="1">  <em>N     </td> <td class="gt_row gt_center" rowspan="1" colspan="2"> 5381   </td> </tr> <tr>
#>     <td class="gt_row gt_left" rowspan="1" colspan="1">  <em>R<sup>2</sup>     </td> <td class="gt_row gt_center" rowspan="1" colspan="2"> .02   </td> </tr> </tbody>
#>   <tfoot class="gt_sourcenotes">
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3"><span class='gt_from_md'>Given that dummy variables lose their interpretability when standardized, only continuous predictors are standardized.</span></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_sourcenote" colspan="3"><span class='gt_from_md'>&dagger; <em>p</em> &lt; .1, * <em>p</em> &lt; .05, ** <em>p</em> &lt; .01, *** <em>p</em> &lt; .001</span></td>
#>     </tr>
#>   </tfoot>
#>   
#> </table>
#> </div>
#> </body>
#> </html>
#> 

  

      	OR [95% CI]
      	Stand. OR [95% CI]
    

religionyes
	0.95     [0.82, 1.09]
	0.95     [0.82, 1.09]

age
	1.01 *** [1.01, 1.02]
	1.25 *** [1.19, 1.32]

gendermale
	1.16 **  [1.05, 1.29]
	1.16 **  [1.05, 1.29]

  N      	 5381    

  R2 	 .02    

Given that dummy variables lose their interpretability when standardized, only continuous predictors are standardized.
    

† p < .1, * p < .05, ** p < .01, *** p < .001