Create a summary table comparing standardized and non-standardized proportional odd logistic regression models

This function creates a summary table for polr models (including mice::mira objects containing polr-models) that shows a standardized and non-standardized version of the model side-by-side. Several pairs of such models can be compared side-by-side.

report_polr_with_std(
  mod,
  mod_std,
  OR = TRUE,
  conf_level = 0.95,
  fmt = "%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  notes = list(),
  apa_style = TRUE,
  stars = std_stars_pad,
  ...
)

Arguments

mod

A polr-model/mira object of polr models, with variables not standardized (or a list of such models)

mod_std

A polr-model/mira object of polr models, with standardized predictor variables (or a list of such models)

OR

Logical. Should odds ratios be shown instead of typical coefficients. If TRUE, estimates are exponentiated

conf_level

Confidence level to use for confidence intervals, defaults to .95

fmt

determines how to format numeric values

integer: the number of digits to keep after the period format(round(x, fmt), nsmall=fmt)

character: passed to the sprintf function (e.g., '%.3f' keeps 3 digits with trailing zero). See ?sprintf

function: returns a formatted character string.

NULL: does not format numbers, which allows users to include function in the "glue" strings in the estimate and statistic arguments.

A named list to format distinct elements of the table differently. Names correspond to column names produced by get_estimates(model) or get_gof(model). Values are integers, characters, or functions, as described above. The fmt element is used as default for unspecified elements Ex: fmt=list("estimate"=2, "std.error"=1, "r.squared"=4, "fmt"=3)

LaTeX output: To ensure proper typography, all numeric entries are enclosed in the \num{} command, which requires the siunitx package to be loaded in the LaTeX preamble. This behavior can be altered with global options. See the 'Details' section.

statistic_vertical

Should standard errors and CIs be shown below coefficients? Defaults to horizontal layout

filename

the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)

model_names

If several pairs of models are to be plotted side by side, indicate the label for each *pair* here

show_nimp

Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic

notes

List of notes to append to bottom of table. An explanation of significance stars is automatically added. A note is also added stating that dummy variables were not scaled in standardization. If you approached standardisation differently, that should be removed.

apa_style

Logical, should APA-style formatting be applied

stars

Named vector of significance stars and their thresholds, check `timesaveR:::std_stars_pad` for default.

...

Arguments passed on to modelsummary::modelsummary

output

filename or object type (character string)

Supported filename extensions: .docx, .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "gt", "kableExtra", "huxtable", "flextable", "jupyter". The "modelsummary_list" value produces a lightweight object which can be saved and fed back to the modelsummary function.

Warning: Users should not supply a file name to the output argument if they intend to customize the table with external packages. See the 'Details' section.

LaTeX compilation requires the booktabs and siunitx packages, but siunitx can be disabled or replaced with global options. See the 'Details' section.

The default output formats and table-making packages can be modified with global options. See the 'Details' section.

estimate

a single string or a character vector of length equal to the number of models. Valid entries include any column name of the data.frame produced by get_estimates(model), and strings with curly braces compatible with the glue package format. Examples:

"estimate"

"{estimate} ({std.error}){stars}"

"{estimate} [{conf.low}, {conf.high}]"

vcov

robust standard errors and other manual statistics. The vcov argument accepts six types of input (see the 'Details' and 'Examples' sections below):

NULL returns the default uncertainty estimates of the model object

string, vector, or (named) list of strings. "iid", "classical", and "constant" are aliases for NULL, which returns the model's default uncertainty estimates. The strings "HC", "HC0", "HC1" (alias: "stata"), "HC2", "HC3" (alias: "robust"), "HC4", "HC4m", "HC5", "HAC", "NeweyWest", "Andrews", "panel-corrected", "outer-product", and "weave" use variance-covariance matrices computed using functions from the sandwich package, or equivalent method. The behavior of those functions can (and sometimes must) be altered by passing arguments to sandwich directly from modelsummary through the ellipsis (...), but it is safer to define your own custom functions as described in the next bullet.

function or (named) list of functions which return variance-covariance matrices with row and column names equal to the names of your coefficient estimates (e.g., stats::vcov, sandwich::vcovHC, function(x) vcovPC(x, cluster="country")).

formula or (named) list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).

named list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a named list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below. Warning: since this list of vectors can include arbitrary strings or numbers, modelsummary cannot automatically calculate p values. The stars argument may thus use incorrect significance thresholds when vcov is a list of vectors.

exponentiate

TRUE, FALSE, or logical vector of length equal to the number of models. If TRUE, the estimate, conf.low, and conf.high statistics are exponentiated, and the std.error is transformed to exp(estimate)*std.error.

shape

formula which determines the shape of the table. The left side determines what appears on rows, and the right side determines what appears on columns. The formula can include a group identifier to display related terms together, which can be useful for models with multivariate outcomes or grouped coefficients (See examples section below). This identifier must be one of the column names produced by: get_estimates(model). The group identifier can be combined with the term identifier in a single column by using the colon to represent an interaction. If an incomplete formula is supplied (e.g., ~statistic), modelsummary tries to complete it automatically. Potential shape values include:

term + statistic ~ model: default

term ~ model + statistic: statistics in separate columns

model + statistic ~ term: models in rows and terms in columns

term + response + statistic ~ model: term and group id in separate columns

term : response + statistic ~ model: term and group id in a single column

term ~ response

coef_map

character vector. Subset, rename, and reorder coefficients. Coefficients omitted from this vector are omitted from the table. The order of the vector determines the order of the table. coef_map can be a named or an unnamed character vector. If coef_map is a named vector, its values define the labels that must appear in the table, and its names identify the original term names stored in the model object: c("hp:mpg"="HPxM/G"). See Examples section below.

coef_omit

string regular expression (perl-compatible) used to determine which coefficients to omit from the table. A "negative lookahead" can be used to specify which coefficients to keep in the table. Examples:

"ei": omit coefficients matching the "ei" substring.

"^Volume$": omit the "Volume" coefficient.

"ei|rc": omit coefficients matching either the "ei" or the "rc" substrings.

"^(?!Vol)": keep coefficients starting with "Vol" (inverse match using a negative lookahead).

"^(?!.*ei)": keep coefficients matching the "ei" substring.

"^(?!.*ei|.*pt)": keep coefficients matching either the "ei" or the "pt" substrings.

See the Examples section below for complete code.

coef_rename

named character vector or function

Named character vector: Values refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object. Ex: c("hp:mpg"="hp X mpg")

Function: Accepts a character vector of the model's term names and returns a named vector like the one described above. The modelsummary package supplies a coef_rename() function which can do common cleaning tasks: modelsummary(model, coef_rename = coef_rename)

gof_map

rename, reorder, and omit goodness-of-fit statistics and other model information. This argument accepts 4 types of values:

NULL (default): the modelsummary::gof_map dictionary is used for formatting, and all unknown statistic are included.

character vector: "all", "none", or a vector of statistics such as c("rmse", "nobs", "r.squared"). Elements correspond to colnames in the data.frame produced by get_gof(model). The modelsummary::gof_mapdefault dictionary is used to format and rename statistics.

NA: excludes all statistics from the bottom part of the table.

data.frame with 3 columns named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples' section below.

list of lists, each of which includes 3 elements named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples section below'.

gof_omit

string regular expression (perl-compatible) used to determine which statistics to omit from the bottom section of the table. A "negative lookahead" can be used to specify which statistics to keep in the table. Examples:

"IC": omit statistics matching the "IC" substring.

"BIC|AIC": omit statistics matching the "AIC" or "BIC" substrings.

"^(?!.*IC)": keep statistics matching the "IC" substring.

group_map

named or unnamed character vector. Subset, rename, and reorder coefficient groups specified a grouping variable specified in the shape argument formula. This argument behaves like coef_map.

add_rows

a data.frame (or tibble) with the same number of columns as your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the row positions. See Examples section below.

align

A string with a number of characters equal to the number of columns in the table (e.g., align = "lcc"). Valid characters: l, c, r, d.

"l": left-aligned column

"c": centered column

"r": right-aligned column

"d": dot-aligned column. Only supported for LaTeX/PDF tables produced by kableExtra. These commands must appear in the LaTeX preamble (they are added automatically when compiling Rmarkdown documents to PDF):

\usepackage{booktabs}

\usepackage{siunitx}

\newcolumntype{d}{S[input-symbols = ()]}

title

string

escape

boolean TRUE escapes or substitutes LaTeX/HTML characters which could prevent the file from compiling/displaying. This setting does not affect captions or notes.

Details

The R2 shown is the maximum likelihood pseudo R2 returned by pR2.

Examples

library(MASS)
#> 
#> Attaching package: ‘MASS’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
pov_att <- polr(poverty ~ religion + age + gender, data = WVS)
pov_att_std <- polr_std(poverty ~ religion + age + gender, data = WVS)
report_polr_with_std(pov_att, pov_att_std, coef_omit = "\\|")
#> fitting null model for pseudo-r2